# perf(uuid): make `bytesToUuid()` up to 100x faster